librumpxen_xendev: xenbus: Reorganise to split minios from rumpkernel parts
Split the xenbus driver into two pieces, busdev.c (in the netbsd
kernel namespacve) and busdev_user.c (in the minios namespace).  They
communicate via the (somewhat ad-hoc) interface in busdev_user.h.  The
interface uses `rumpxenbus_*' names so that the two sides can call
each other.  We split the state structure up into three: a netbsd
part, a minios part, and a common part.

This is actually largely a combination of code motion and function and
type renaming.  There is little functional change from the previous
"header abuse" approach, other than some minor interface adjustments.

In much of the code `d' was used to refer to the device struct.  Now
there are three context stucts.  For now I have retained in each
function the use of `d' which producese the lowest amount of code
churn.  The compiler makes sure the types all line up right.

Later, the uses of `d' will each be changed to `dc', `du' or `dd'.
