+++ platform/xen/librumpxen_xendev/Makefile
@@ -1,3 +1,5 @@
+.include <bsd.own.mk>
+
 LIB=	rumpxen_xendev
 
 SRCS=	xendev_component.c
@@ -9,16 +11,16 @@ RUMPTOP= ${TOPRUMP}
 
 CPPFLAGS+=	-I${RUMPTOP}/librump/rumpkern -I${RUMPTOP}/librump
 CPPFLAGS+=	-I${RUMPTOP}/librump/rumpvfs
+CPPFLAGS+=	-I${.CURDIR}/../xen/include
 CPPFLAGS+=	-I${.CURDIR}
-CPPFLAGS+=	-I${.CURDIR}/../xen/include -D__RUMP_KERNEL__ -I${.CURDIR}/..
-CPPFLAGS+=	-I${.CURDIR}/../../../include
 
-.if ${BUILDRR:Uno} == "true"
-.include "${RUMPRUN_MKCONF}"
-CPPFLAGS+=	-I${OBJDIR}/dest.stage/include
-.endif
+RUMPCOMP_USER_SRCS=	busdev_user.c
+RUMPCOMP_USER_CPPFLAGS+= -I${.CURDIR}/..
+RUMPCOMP_USER_CPPFLAGS+= -I${.CURDIR}/../xen/include
+RUMPCOMP_USER_CPPFLAGS+= -I${.CURDIR}/../../../include
 
-RUMP_SYM_NORENAME=xenbus_|HYPERVISOR_|minios_|bmk_
+# XXX
+.undef RUMPKERN_ONLY
 
 .include "${RUMPTOP}/Makefile.rump"
 .include <bsd.lib.mk>
