+++ platform/xen/librumpxen_xendev/rumpxen_xendev.h
@@ -39,11 +39,6 @@
 #include <sys/poll.h>
 
 
-/* nicked from NetBSD sys/dev/pci/cxgb/cxgb_adapter.h */
-#ifndef container_of
-#define container_of(p, stype, field) ((stype *)(((uint8_t *)(p)) - offsetof(stype, field)))
-#endif
-
 //#define RUMP_DEV_XEN_DEBUG 1
 
 #ifdef RUMP_DEV_XEN_DEBUG
